+++ Scripts/Constructables/Buildings/BuildingResourceEndpoint.cs
@@ -21,22 +21,22 @@ public sealed class BuildingResourceEndpoint : IResourceEndpoint
 
     public Building Owner => _owner;
 
-    public float DepositResource(string resourceId, float amount)
+    public int DepositResource(string resourceId, float amount)
     {
         return _owner.BulkStorage.Deposit(resourceId, amount);
     }
 
-    public float WithdrawResource(string resourceId, float amount)
+    public int WithdrawResource(string resourceId, int amount)
     {
         return _owner.BulkStorage.Withdraw(resourceId, amount);
     }
 
-    public float GetStockpile(string resourceId)
+    public int GetStockpile(string resourceId)
     {
         return _owner.BulkStorage.GetQuantity(resourceId);
     }
 
-    public IReadOnlyDictionary<string, float> GetAllStockpiles()
+    public IReadOnlyDictionary<string, int> GetAllStockpiles()
     {
         return _owner.BulkStorage.GetAllQuantities();
     }
@@ -56,13 +56,13 @@ public sealed class BuildingResourceEndpoint : IResourceEndpoint
         if (bulk == null || bulk.Slots.Count == 0)
             return 0f;
 
-        float used = 0f;
-        float capacity = 0f;
+        int used = 0;
+        int capacity = 0;
         foreach (var slot in bulk.Slots)
         {
             used += slot.Quantity;
             capacity += slot.Capacity;
         }
-        return capacity > 0f ? used / capacity : 0f;
+        return capacity > 0 ? (float)used / capacity : 0f;
     }
 }
