+++ Scripts/Constructables/Stations/StationResourceEndpoint.cs
@@ -18,16 +18,16 @@ public sealed class StationResourceEndpoint : IResourceEndpoint
 
     public StationSatellite Owner => _owner;
 
-    public float DepositResource(string resourceId, float amount)
+    public int DepositResource(string resourceId, float amount)
         => _owner.BulkStorage.Deposit(resourceId, amount);
 
-    public float WithdrawResource(string resourceId, float amount)
+    public int WithdrawResource(string resourceId, int amount)
         => _owner.BulkStorage.Withdraw(resourceId, amount);
 
-    public float GetStockpile(string resourceId)
+    public int GetStockpile(string resourceId)
         => _owner.BulkStorage.GetQuantity(resourceId);
 
-    public IReadOnlyDictionary<string, float> GetAllStockpiles()
+    public IReadOnlyDictionary<string, int> GetAllStockpiles()
         => _owner.BulkStorage.GetAllQuantities();
 
     public void EnqueueResourceRequest(ResourceRequest request) { }
@@ -37,8 +37,8 @@ public sealed class StationResourceEndpoint : IResourceEndpoint
         if (owner != _owner) return 0f;
         var bulk = _owner.BulkStorage;
         if (bulk.Slots.Count == 0) return 0f;
-        float used = 0f, capacity = 0f;
+        int used = 0, capacity = 0;
         foreach (var slot in bulk.Slots) { used += slot.Quantity; capacity += slot.Capacity; }
-        return capacity > 0f ? used / capacity : 0f;
+        return capacity > 0 ? (float)used / capacity : 0f;
     }
 }
