+++ Scripts/Structures/GameState/IResourceEndpoint.cs
@@ -6,30 +6,32 @@ namespace Structures.GameState;
 /// <summary>
 /// Common interface for entities that can send and receive resources via transfers.
 /// Implemented by ContinentEconomy and StationEconomy.
+/// Resources are accounted as whole units; production paths may pass fractional
+/// deposits which the underlying storage buffers internally.
 /// </summary>
 public interface IResourceEndpoint
 {
     /// <summary>
-    /// Deposits resources into this endpoint's stockpile.
-    /// Returns the amount actually deposited (may be less than requested due to capacity).
+    /// Deposits resources into this endpoint's stockpile. Accepts a fractional amount;
+    /// returns whole units actually placed (sub-unit residue is buffered internally).
     /// </summary>
-    float DepositResource(string resourceId, float amount);
+    int DepositResource(string resourceId, float amount);
 
     /// <summary>
-    /// Withdraws resources from this endpoint's stockpile.
+    /// Withdraws whole resource units from this endpoint's stockpile.
     /// Returns the amount actually withdrawn (may be less than requested due to availability).
     /// </summary>
-    float WithdrawResource(string resourceId, float amount);
+    int WithdrawResource(string resourceId, int amount);
 
     /// <summary>
-    /// Gets the current stockpile quantity of a specific resource.
+    /// Gets the current stockpile quantity of a specific resource (whole units).
     /// </summary>
-    float GetStockpile(string resourceId);
+    int GetStockpile(string resourceId);
 
     /// <summary>
-    /// Gets all current stockpile quantities.
+    /// Gets all current stockpile quantities (whole units).
     /// </summary>
-    IReadOnlyDictionary<string, float> GetAllStockpiles();
+    IReadOnlyDictionary<string, int> GetAllStockpiles();
 
     /// <summary>
     /// Queues a resource request for discrete manufacturing.
